refactor: 친구 요청보내기 Service로직 예외 사항 `validateSendRequest` 메서드로 분리 및 Collectors.toList() → stream().toList() 로 변경
