+++ src/main/java/com/samsamhajo/deepground/friend/Dto/FriendRequestDto.java
@@ -4,9 +4,11 @@ import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-@Data
-@AllArgsConstructor
+@Getter
+@NoArgsConstructor
 public class FriendRequestDto {
 
         private Long requesterId;
@@ -14,5 +16,6 @@ public class FriendRequestDto {
         @NotBlank(message = "이메일을 입력해주세요")
         @Email(message = "올바른 이메일 형식을 입력해주세요")
         private String receiverEmail;
-    }
+
+}
